show time mode selector in vessel profile and report
